--- src/main/java/edu/ucsb/cs156/example/controllers/ArticlesContoller.java
@@ -1,78 +1,0 @@
-import edu.ucsb.cs156.example.repositories.UserRepository;
-import edu.ucsb.cs156.example.testconfig.TestConfig;
-import edu.ucsb.cs156.example.ControllerTestCase;
-import edu.ucsb.cs156.example.entities.Articles;
-import edu.ucsb.cs156.example.entities.Articles;
-import edu.ucsb.cs156.example.repositories.ArticlesRepository;
-import edu.ucsb.cs156.example.repositories.ArticlesRepository;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Map;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Import;
-import org.springframework.http.MediaType;
-import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.web.servlet.MvcResult;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
-
-import java.util.Optional;
-
-import javax.management.ConstructorParameters;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-@WebMvcTest(controllers = ArticlesController.class)
-@Import(TestConfig.class)
-public class ArticlesController {
-    @Autowired
-    ArticlesRepository articlesRepository;
-
-    @Operation(summary= "List all articles")
-    @PreAuthorize("hasRole('ROLE_USER')")
-    @GetMapping("/all")
-    public Iterable<Articles> allArticless() {
-        Iterable<Articles> dates = articlesRepository.findAll();
-        return dates;
-    }
-
-    @Operation(summary= "Create a new article")
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @PostMapping("/post")
-    public Articles postArticles(
-            @Parameter(name="title") @RequestParam String title,
-            @Parameter(name="url") @RequestParam String url, 
-            @Parameter(name= "email") @RequestParam String email,
-            @Paramater(name = "explanation") @RequestParam String explanation,
-            @Parameter(name="dateAdded", description="in iso format, e.g. YYYY-mm-ddTHH:MM:SS; see https://en.wikipedia.org/wiki/ISO_8601")  @RequestParam("localDateTime") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime localDateTime)
-            throws JsonProcessingException {
-
-        // For an explanation of @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
-        // See: https://www.baeldung.com/spring-date-parameters
-
-        log.info("localDateTime={}", localDateTime);
-
-        Articles articles = Articles();
-        articles.setTitle(title);
-        articles.setUrl(url);
-        articles.setEmail(email);
-        articles.setExplanation(explanation);
-        articles.setLocaldateAdded(dateAdded);
-
-        Articles savedUcsbDate = articlesRepository.save(articles);
-
-        return savedUcsbDate;
-    }
- 
-}
