+++ src/main/java/edu/ucsb/cs156/example/controllers/ArticlesController.java
@@ -1,0 +1,67 @@
+package edu.ucsb.cs156.example.controllers;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.extern.slf4j.Slf4j;
+import edu.ucsb.cs156.example.entities.Articles;
+
+import edu.ucsb.cs156.example.repositories.ArticlesRepository;
+
+
+import java.time.LocalDateTime;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+@Tag(name = "Articles")
+@RequestMapping("/api/articles")
+@RestController
+@Slf4j
+public class ArticlesController {
+    @Autowired
+    ArticlesRepository articlesRepository;
+
+    @Operation(summary= "List all articles")
+    @PreAuthorize("hasRole('ROLE_USER')")
+    @GetMapping("/all")
+    public Iterable<Articles> allArticless() {
+        Iterable<Articles> dates = articlesRepository.findAll();
+        return dates;
+    }
+
+    @Operation(summary= "Create a new article")
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PostMapping("/post")
+    public Articles postArticles(
+            @Parameter(name="title") @RequestParam String title,
+            @Parameter(name="url") @RequestParam String url, 
+            @Parameter(name= "email") @RequestParam String email,
+            @Parameter(name = "explanation") @RequestParam String explanation,
+            @Parameter(name="dateAdded", description="in iso format, e.g. YYYY-mm-ddTHH:MM:SS; see https://en.wikipedia.org/wiki/ISO_8601")  @RequestParam("dateAdded") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime dateAdded)
+            throws JsonProcessingException {
+
+        // For an explanation of @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
+        // See: https://www.baeldung.com/spring-date-parameters
+
+        log.info("dateAdded={}", dateAdded);
+
+        Articles articles = new Articles();
+        articles.setTitle(title);
+        articles.setUrl(url);
+        articles.setEmail(email);
+        articles.setExplanation(explanation);
+        articles.setDateAdded(dateAdded);
+
+        Articles savedUcsbDate = articlesRepository.save(articles);
+
+        return savedUcsbDate;
+    }
+ 
+}
